+++ up42/__init__.py
@@ -22,14 +22,13 @@ from typing import cast
 
 # pylint: disable=only-importing-modules-is-allowed
 from up42.base import authenticate, stac_client
-from up42.coverage import OrderCoverage
 from up42.glossary import CollectionSorting, CollectionType, ProductGlossary, Provider
 from up42.order import Order, OrderSorting
 from up42.order_template import BatchOrderTemplate
 from up42.processing import Job, JobSorting, JobStatus
 from up42.stac import BulkDeletion
 from up42.stac import extend as stac_extend
-from up42.tasking import FeasibilityStudy, FeasibilityStudySorting, Quotation, QuotationSorting
+from up42.tasking import FeasibilityStudy, FeasibilityStudySorting, OrderCoverage, Quotation, QuotationSorting
 from up42.utils import get_up42_py_version
 from up42.version import version_control