+++ up42/tasking.py
@@ -6,6 +6,8 @@ import dataclasses
 from collections.abc import Iterator
 from typing import Literal, TypeAlias
 
+import geojson  # type: ignore
+
 from up42 import base, host, utils
 
 logger = utils.get_logger(__name__)
@@ -169,3 +171,42 @@ class FeasibilityStudy:
         feasibility_study = self._from_metadata(metadata)
         for field in dataclasses.fields(feasibility_study):
             setattr(self, field.name, getattr(feasibility_study, field.name))
+
+
+@dataclasses.dataclass
+class GeometryMetrics:
+    sq_km_area: float
+    percentage: float
+    geometry: geojson.Polygon | geojson.MultiPolygon
+
+
+@dataclasses.dataclass
+class OrderCoverage:
+    session = base.Session()
+    covered: GeometryMetrics
+    remainder: GeometryMetrics
+
+    @classmethod
+    def get(cls, order_id: str) -> "OrderCoverage":
+        url = host.endpoint(f"/v2/coverage/orders/{order_id}")
+        metadata = cls.session.get(url=url).json()
+        return OrderCoverage._from_metadata(metadata)
+
+    @staticmethod
+    def _from_metadata(metadata: dict) -> "OrderCoverage":
+        covered_data = metadata["covered"]
+        remainder_data = metadata["remainder"]
+
+        covered = GeometryMetrics(
+            sq_km_area=covered_data["sqKmArea"],
+            percentage=covered_data["percentage"],
+            geometry=covered_data["geometry"],
+        )
+
+        remainder = GeometryMetrics(
+            sq_km_area=remainder_data["sqKmArea"],
+            percentage=remainder_data["percentage"],
+            geometry=remainder_data["geometry"],
+        )
+
+        return OrderCoverage(covered=covered, remainder=remainder)